requirements/local.txt
asgiref==3.10.0
Django==5.2.8
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-model-utils==5.0.0
pillow==12.0.0
psycopg2==2.9.11
sqlparse==0.5.3
tzdata==2025.2
Unipath==1.1
